Move mock into its own directory
